refactor: Rename NaverResponse to NaverShoppingResponse
