+++ src/types/NaverShoppingResponse.ts
